Move tests out of cli crate
